added POST and GET for Building
